account method ok

+++ src/main/java/com/myBudget/myBudget/models/Transaction.java
@@ -7,6 +7,7 @@ import javax.persistence.Id;
 import javax.persistence.Temporal;
 import javax.persistence.TemporalType;
 import java.util.Date;
+import java.math.BigDecimal;
 
 @Entity
 public class Transaction {
@@ -19,11 +20,11 @@ public class Transaction {
     private Date date;
 
     private String category;
-    private Integer amount;
+    private BigDecimal amount;
     private Integer accountId;
 
     // Constructor
-    public Transaction(Integer id, Date date, String category, Integer amount, Integer accountId) {
+    public Transaction(Integer id, Date date, String category, BigDecimal amount, Integer accountId) {
         this.id = id;
         this.date = date;
         this.category = category;
@@ -39,14 +40,14 @@ public class Transaction {
     public Integer getId() { return id; }
     public Date getDate() { return date; }
     public String getCategory() { return category; }
-    public Integer getAmount() { return amount; }
+    public BigDecimal getAmount() { return amount; }
     public Integer getAccountId() { return accountId; }
 
     // Setters
     public void setId(Integer id) { this.id = id; }
     public void setDate(Date date) { this.date = date; }
     public void setCategory(String category) { this.category = category; }
-    public void setAmount(Integer amount) { this.amount = amount; }
+    public void setAmount(BigDecimal amount) { this.amount = amount; }
     public void setAccountId(Integer accountId) { this.accountId = accountId; }
 
 }
